manga-organizer/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-generativeai==0.8.3
python-dotenv==1.0.0
